ファイル一覧のレイアウトを修正

+++ src/show.php
@@ -1,6 +1,8 @@
+<?php
+session_start();
+?>
 <!DOCTYPE html>
 <html>
-
 <head>
     <meta charset="UTF-8">
     <link rel="stylesheet" href="../css/style.css">
@@ -11,6 +13,7 @@
 <?php
 include 'menu.php';
 require_once 'get_data.php';
+require_once 'auth.php'; 
 
 $targetGenre = $_GET['genre'] ?? 'すべて';
 
@@ -29,18 +32,18 @@ echo '<div class="sort-select">';
 echo '<select id="sort-select" onchange="changeSort(this.value)" >';
     echo '<option value="new" ' . ($sort == 'new' ? 'selected' : '') . '>新しい順</option>';
     echo '<option value="old" ' . ($sort == 'old' ? 'selected' : '') . '>古い順</option>';
-    echo '<option value="summary_asc" ' . ($sort == 'summary_asc' ? 'selected' : '') . '>説明(昇順)</option>';
-    echo '<option value="summary_desc" ' . ($sort == 'summary_desc' ? 'selected' : '') . '>説明(降順)</option>';
+    //echo '<option value="summary_asc" ' . ($sort == 'summary_asc' ? 'selected' : '') . '>五十音順(昇順)</option>';
+    //echo '<option value="summary_desc" ' . ($sort == 'summary_desc' ? 'selected' : '') . '>五十音順(降順)</option>';
 echo '</select>';
 echo '</div>';
-/*
-echo '<div class="list" style="margin-top:70px;">';
-    echo '<div style="width:40%">ファイル説明</div>';
+
+echo '<div class="list-title" >';
+    echo '<div style="width:30%">ファイル説明</div>';
     echo '<div style="width:20%">ジャンル</div>';
-    echo '<div style="20%"></div>';
-    echo '<div style="20%"></div>';
+    echo '<div style="width:20%">日時</div>';
+    echo '<div style="30%">　</div>';
 echo '</div>';
-*/
+
 echo '<hr>';
 
 //fileの中のpdfのみ探す
@@ -59,6 +62,7 @@ foreach($csvData as $i => $file_path){
 
     $downloadName = $fileSummary.'.pdf';
     $filePath = "../file/" . $fileName;
+    $fileView = '../file/' . $info['name'];
     
     //すべて表示でないときとジャンルが一致しないときはスキップ
     if($targetGenre !== 'すべて' && $fileGenre !== $targetGenre){
@@ -66,7 +70,8 @@ foreach($csvData as $i => $file_path){
     }
 
     echo '<div class="list">';
-        echo '<div style="width:30%">' . htmlspecialchars($fileSummary). '</div>'; // 説明
+        echo '<div style="width:25%">' . htmlspecialchars($fileSummary).'</div>'; // 説明
+        echo '<div style="width:5%"><a href="' . $filePath . '">確認</a></div>';
         echo '<div style="width:20%">' . htmlspecialchars($fileGenre) . '</div>'; // ジャンル
         echo '<div style="width:20%">' . date('Y/m/d', strtotime($fileTime)) . '</div>'; 
 
@@ -74,12 +79,14 @@ foreach($csvData as $i => $file_path){
         echo '<a href="' . $filePath . '" download="' . htmlspecialchars($downloadName). '" class="btn-download">ダウンロード</a>';
         echo '</div>';
         
-        echo '<div style="width:10%" >';
-        echo '<form action="delete.php" method="post" class="delete-form">';
-        echo '<input type="hidden" name="delete" value="' . $fileName . '">';
-        echo '<button type="submit" onclick="return confirm(\'' . $fileSummary . 'を削除しますか？\')" class="btn-delete">削除</button>';
-        echo '</form>';
-        echo '</div>';
+        if (isAdmin()) {
+            echo '<div style="width:10%" >';
+            echo '<form action="delete.php" method="post" class="delete-form">';
+            echo '<input type="hidden" name="delete" value="' . $fileName . '">';
+            echo '<button type="submit" onclick="return confirm(\'' . $fileSummary . 'を削除しますか？\')" class="btn-delete">削除</button>';
+            echo '</form>';
+            echo '</div>';
+        }
     echo '</div>';
     echo '<hr>';
 }
